+++ include/OthelloAI.h
@@ -3,13 +3,13 @@
 class StoneScoreList {
   int hand_x;
   int hand_y;
-  std::vector<double> score;
-  double total_score;
+  std::vector<int> score;
+  int total_score;
 public:
   StoneScoreList(int x, int y);
   void set_total_score();
   void get_coordinate(int &x, int &y);
-  void set_score(double s);
+  void set_score(int s);
   void show_score_list();
   bool operator>(const StoneScoreList &right)const;
   inline bool is_edge(int x, int y);
@@ -34,5 +34,5 @@ class OthelloAI {
   void seek(int max_depth);
   void record_dot_stone();
   void set_subAI(int branch);
-  double get_avarage_score();
+  long long get_avarage_score();
 };
